Feat. 레디스 설정 변경 & 대기열 redis로 로직 변경 초안

+++ src/main/java/com/hh99/hh5concertreservation/waiting/interfaces/presentation/dto/WaitingOrderResponse.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class WaitingOrderResponse {
     private Integer status;
-    private Long waitingCount;
+    private Integer waitingCount;
 
     public WaitingOrderResponse(CheckStateResult checkStateResult) {
         this.status = checkStateResult.getStatus();
